+++ PinnyNotes.DataAccess/DatabaseInitializer.cs
@@ -26,6 +26,10 @@ public class DatabaseInitialiser
                 INSERT INTO SchemaInfo (Id, Version) VALUES
                     (0, @schemaVersion);
 
+                CREATE TABLE IF NOT EXISTS {ApplicationDataRepository.TableName}
+                    {ApplicationDataRepository.TableSchema};
+                INSERT INTO {ApplicationDataRepository.TableName} DEFAULT VALUES;
+
                 CREATE TABLE IF NOT EXISTS {SettingsRepository.TableName}
                     {SettingsRepository.TableSchema};
                 INSERT INTO {SettingsRepository.TableName} DEFAULT VALUES;